Small refinements and improved arb transaction lookup
As we can not rely on the state of the RPC when it is running the trans lookup, if it is behind it can give empty results. Delay logic is in place but that appears to not be enough. I added a backup to pull the main Solana RPC. Should be very light load as we are only checking one transaction at a time and only on successful ARB. This is a good fall back for the default RPC set by the user and hopefully solve issues when the RPC does not return the transaction status within the retry / pause timeframe. Of course the backup can be easily changed for PRO users in the ./src/utils/transaction.js file.